Manage JSON collections using MySQL and/or PostgreSQL
We add the possibility to store objects instantiated from models which
embed `kolektor.Model` and implement the `Modeler.CollectionName()`
function, into SQL tables using an opinionated but schema less DDL.

For PostgreSQL we load the UUID functionality, and make sure that the
`uid` field is set to an UUIDv4 when inserting data.
For MySQL, a stored function is added which generates a UUIDv4, and
does the same as for PostgreSQL.

For example, to store books into PostgreSQL:

    session, _ := NewSession(stores.PgSQL, dsn)
    books, err := session.Collection(&Book{})
    book := &Book{
        ISBN13: "978-0135800911",
        Title:  "Presentation Zen",
    }
    books.Store(book)

    book := &Book{}
    _ := books.Get(book, book.Meta.ID)

The example above is exactly the same for MySQL, except using
`stores.MySQL` and different Data Source Name (DSN).

The `Meta` attribute for models contains `ID` (data store primary key),
the `UID` (Unique ID) which for now is a UUIDv4, and the timestamps
`Created` and `Updated` which respectively show when the object was
first stored and when it was updated.

Build tags have been added, making it possible to disable data stores.
For example, `go build -tags nopgsql`, will disable compiling PostgreSQL
support.

Tests have been added for both MySQL and PostgreSQL. It is best to start
both RDBMS using Docker Compose in `_support/docker-compose`.