Change GetObject() allowing selection using multiple fields
We change the store's `GetObject()`-method so that it accepts a mapping
of field/value instead of only one. This makes it possible to retrieve
objects based on 1 or more attributes.

The Collection methods did not change, so there is no breaking change
for users. We do however add a new method `Collection.GetByFields` so
that it is possible to use multiple fields instead of the Get-method
which only allows to use (secondary) primary keys.